moves sselvish to separate file
